0.17.x into develop

+++ kalite/updates/videos.py
@@ -10,6 +10,7 @@ from . import settings
 from django.conf import settings as django_settings
 from fle_utils.general import ensure_dir
 from fle_utils.internet.download import callback_percent_proxy, download_file
+from requests.exceptions import HTTPError
 
 
 logger = logging.getLogger(__name__)
@@ -18,8 +19,8 @@ logger = logging.getLogger(__name__)
 def get_local_video_size(youtube_id, default=None):
     try:
         return os.path.getsize(os.path.join(django_settings.CONTENT_ROOT, "%s.mp4" % youtube_id))
-    except Exception as e:
-        logger.debug(str(e))
+    except (IOError, OSError) as e:
+        logger.exception(e)
         return default
 
 
@@ -79,26 +80,36 @@ def download_video(youtube_id, extension="mp4", callback=None):
     filepath = get_video_local_path(youtube_id, extension)
     thumb_filepath = get_thumbnail_local_path(youtube_id)
 
+    logger.info(
+        "Downloading {id} (video: {video}, thumbnail: {thumbnail})".format(
+            id=youtube_id,
+            video=url,
+            thumbnail=thumb_url,
+        )
+    )
+
+    # Download video
     try:
-        # Download video
         download_file(
             url,
             dst=filepath,
             callback=callback_percent_proxy(callback, end_percent=95),
             max_retries=settings.DOWNLOAD_MAX_RETRIES
         )
-
-        # Download thumbnail
-        download_file(
-            thumb_url,
-            dst=thumb_filepath,
-            callback=callback_percent_proxy(callback, start_percent=95, end_percent=100),
-            max_retries=settings.DOWNLOAD_MAX_RETRIES
+    except HTTPError as e:
+        logger.error(
+            "HTTP status {status} for URL: {url}".format(
+                status=e.response.status_code,
+                url=e.response.url,
+            )
         )
+        raise
 
     except (socket.timeout, IOError) as e:
-        logger.exception(e)
-        logger.info("Timeout -- Network UnReachable")
+        logger.error("Network error for URL: {url}, exception: {exc}".format(
+            url=url,
+            exc=str(e)
+        ))
         delete_downloaded_files(youtube_id)
         raise
 
@@ -107,15 +118,35 @@ def download_video(youtube_id, extension="mp4", callback=None):
         delete_downloaded_files(youtube_id)
         raise
 
+    # Download thumbnail - don't fail if it doesn't succeed, because at this
+    # stage, we know that the video has been downloaded.
+    try:
+        download_file(
+            thumb_url,
+            dst=thumb_filepath,
+            callback=callback_percent_proxy(callback, start_percent=95, end_percent=100),
+            max_retries=settings.DOWNLOAD_MAX_RETRIES
+        )
+    except HTTPError as e:
+        logger.error(
+            "HTTP status {status} for URL: {url}".format(
+                status=e.response.status_code,
+                url=e.response.url,
+            )
+        )
+    except (socket.timeout, IOError) as e:
+        logger.error("Network error for URL: {url}, exception: {exc}".format(
+            url=url,
+            exc=str(e)
+        ))
+
 
 def delete_downloaded_files(youtube_id):
     download_path = django_settings.CONTENT_ROOT
     files_deleted = 0
     for filepath in glob.glob(os.path.join(download_path, youtube_id + ".*")):
-        try:
+        if os.path.isfile(filepath):
             os.remove(filepath)
             files_deleted += 1
-        except OSError:
-            pass
     if files_deleted:
         return True
